+++ compiler/rustc_trait_selection/src/error_reporting/traits/mod.rs
@@ -2,6 +2,7 @@ pub mod ambiguity;
 pub mod call_kind;
 mod fulfillment_errors;
 pub mod on_unimplemented;
+pub mod on_unimplemented_format;
 mod overflow;
 pub mod suggestions;
